--- a/.gitignore
+++ b/.gitignore
@@ -257,3 +257,4 @@ paket-files/
 
 /tests/CodeCoverage/OpenCover.*
 SixLabors.Shapes.Coverage.xml
+/tests/SixLabors.Shapes.Benchmarks/BenchmarkDotNet.Artifacts/results/