--- a/samples/DrawShapesWithImageSharp/DrawShapesWithImageSharp.csproj
+++ b/samples/DrawShapesWithImageSharp/DrawShapesWithImageSharp.csproj
@@ -15,8 +15,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="ImageSharp.Drawing" Version="1.0.0-alpha2-00115" />
-    <PackageReference Include="ImageSharp.Formats.Png" Version="1.0.0-alpha2-00111" />
+    <PackageReference Include="ImageSharp.Drawing" Version="1.0.0-alpha5-00040" />
+    <PackageReference Include="ImageSharp" Version="1.0.0-alpha5-00046" />
   </ItemGroup>
 
 </Project>