--- a/samples/DrawShapesWithImageSharp/Program.cs
+++ b/samples/DrawShapesWithImageSharp/Program.cs
@@ -166,16 +166,13 @@ namespace SixLabors.Shapes.DrawShapesWithImageSharp
                 img.Fill(Color.DarkBlue);
 
                 // In ImageSharp.Drawing.Paths there is an extension method that takes in an IShape directly.
-                img.Fill(Color.HotPink, new ShapeRegion(shape), new ImageSharp.Drawing.GraphicsOptions(true));
+                img.Fill(Color.HotPink, shape, new ImageSharp.Drawing.GraphicsOptions(true));
                 // img.Draw(Color.LawnGreen, 1, new ShapePath(shape));
 
                 // Ensure directory exists
                 Directory.CreateDirectory(System.IO.Path.GetDirectoryName(fullPath));
 
-                using (var fs = File.Create(fullPath))
-                {
-                    img.SaveAsPng(fs);
-                }
+                    img.Save(fullPath);
             }
         }
 
@@ -188,16 +185,13 @@ namespace SixLabors.Shapes.DrawShapesWithImageSharp
                 img.Fill(Color.DarkBlue);
 
                 // In ImageSharp.Drawing.Paths there is an extension method that takes in an IShape directly.
-                img.Fill(Color.HotPink, new ShapeRegion(shape), new ImageSharp.Drawing.GraphicsOptions(true));
+                img.Fill(Color.HotPink, shape, new ImageSharp.Drawing.GraphicsOptions(true) { AntialiasSubpixelDepth = 0, Antialias = false });
                 // img.Draw(Color.LawnGreen, 1, new ShapePath(shape));
 
                 // Ensure directory exists
                 Directory.CreateDirectory(System.IO.Path.GetDirectoryName(fullPath));
 
-                using (var fs = File.Create(fullPath))
-                {
-                    img.SaveAsPng(fs);
-                }
+                    img.Save(fullPath);
             }
         }
     }