--- a/tests/SixLabors.Shapes.Benchmarks/InteralPath_FindIntersections.cs
+++ b/tests/SixLabors.Shapes.Benchmarks/InteralPath_FindIntersections.cs
@@ -14,7 +14,7 @@ namespace SixLabors.Shapes.Benchmarks
 
         public InteralPath_FindIntersections()
         {
-            this.vectors = new Ellipse(new System.Numerics.Vector2(0, 0), new Size(20, 10))
+            this.vectors = new Ellipse(new System.Numerics.Vector2(0, 0), new Size(2000, 1000))
                 .Flatten()
                 .First().Points.ToArray();
         }
@@ -23,14 +23,11 @@ namespace SixLabors.Shapes.Benchmarks
         public Vector2[] InternalOld()
         {
             Vector2[] buffer = new Vector2[vectors.Length];
-            for (var i = 0; i < 100; i++)
-            {
-                var path = new InternalPath_Old(vectors, true);
+            var path = new InternalPath_Old(vectors, true);
 
-                for (var y = path.Bounds.Top; y < path.Bounds.Bottom; y += (1f / 32f))
-                {
-                    path.FindIntersections(new Vector2(path.Bounds.Left - 1, y), new Vector2(path.Bounds.Right + 1, y), buffer, path.PointCount, 0);
-                }
+            for (var y = path.Bounds.Top; y < path.Bounds.Bottom; y += (1f / 32f))
+            {
+                path.FindIntersections(new Vector2(path.Bounds.Left - 1, y), new Vector2(path.Bounds.Right + 1, y), buffer, path.PointCount, 0);
             }
             return buffer;
         }
@@ -39,14 +36,11 @@ namespace SixLabors.Shapes.Benchmarks
         public Vector2[] InternalNew()
         {
             Vector2[] buffer = new Vector2[vectors.Length];
-            for (var i = 0; i < 100; i++)
-            {
-                var path = new InternalPath(vectors, true);
+            var path = new InternalPath(vectors, true);
 
-                for (var y = path.Bounds.Top; y < path.Bounds.Bottom; y += (1f / 32f))
-                {
-                    path.FindIntersections(new Vector2(path.Bounds.Left - 1, y), new Vector2(path.Bounds.Right + 1, y), buffer, path.PointCount, 0);
-                }
+            for (var y = path.Bounds.Top; y < path.Bounds.Bottom; y += (1f / 32f))
+            {
+                path.FindIntersections(new Vector2(path.Bounds.Left - 1, y), new Vector2(path.Bounds.Right + 1, y), buffer, path.PointCount, 0);
             }
             return buffer;
         }