--- a/tests/SixLabors.Shapes.Benchmarks/Program.cs
+++ b/tests/SixLabors.Shapes.Benchmarks/Program.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Reflection;
+using BenchmarkDotNet.Running;
+
+namespace SixLabors.Shapes.Benchmarks
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            new BenchmarkSwitcher(typeof(Program).GetTypeInfo().Assembly).Run(args);
+            Console.ReadKey();
+        }
+    }
+}