… samples/DrawShapesWithImageSharp/ShapePath.cs
… samples/DrawShapesWithImageSharp/ShapeRegion.cs
… src/SixLabors.Shapes/InternalPath_Old.cs
… tests/SixLabors.Shapes.Benchmarks/InternalPathExperiments/InternalPath_Old.cs
… tests/SixLabors.Shapes.Benchmarks/InternalPathExperiments/InternalPath_Proposal1.cs
… tests/SixLabors.Shapes.Benchmarks/SixLabors.Shapes.Benchmarks.csproj
… tests/SixLabors.Shapes.Tests/GeneralPolygonIntersectionTests.cs
… tests/SixLabors.Shapes.Tests/InternalPathTests.cs
… tests/SixLabors.Shapes.Tests/PathExtensions.cs
… tests/SixLabors.Shapes.Tests/StarTests.cs