Improve the performance of `FindIntersections()` & thus all related operations.

we do 2 major things : 
1. we prefilter the flattened points to remove duplicates and fix long lines bisected by points.
2. we precalculate potentional intersections on weather edges are co-linear to the search line before finding intersections, which lets up drastically reduce the number of times we calculate the orientation between points.

### Some benchmarks
 |      Method |        Mean |    StdDev | Scaled | Scaled-StdDev |
 |------------ |------------ |---------- |------- |-------------- |
 | Old | 255.6094 ms | 6.9209 ms |   1.00 |          0.00 |
 | New |  88.0969 ms | 1.1997 ms |   0.34 |          0.01 |